% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges2faces.R
\name{edges2faces}
\alias{edges2faces}
\title{edges2faces}
\usage{
edges2faces(edges)
}
\arguments{
\item{edges}{numeric matrix with two columns: index of points which have a edge}
}
\value{
faces numeric matrix with three columns: index of points which form a triangle
}
\description{
Given a set of edges between vertices (with xy coordinates) it returns a matrix of all triangles formed by the edges.
}
\examples{
edges <- cbind(c(1,1,1,2,2,3), c(2,3,4,3,4,4))
edges2faces(edges)
}
