% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(g)
}
\arguments{
\item{g}{a web}
}
\value{
an error value
}
\description{
For given web and a portal set it computes the number of intersections of links and the number of portals in each field.
These two numbers give the error of the web.
}
\examples{
xy <- cbind(runif(30), runif(30))
g  <- fishbone(8)
evaluate(g)
gf <- web(g, xy)
evaluate(gf)
}
