% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluateC}
\alias{evaluateC}
\title{Compute the number line intersections for a given web}
\usage{
evaluateC(g)
}
\arguments{
\item{g}{a web}
}
\value{
a vector with number of intersections for each line segment (Ingress: link)
}
\description{
Compute the number line intersections for a given web
}
\references{
\url{https://stackoverflow.com/questions/20519431/finding-point-of-intersection-in-r}
}
