% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbone.R
\name{fishbone}
\alias{fishbone}
\title{fishbone}
\usage{
fishbone(n)
}
\arguments{
\item{n}{numeric: number of links at base portals}
}
\value{
a fishbone web
}
\description{
Creates a fishbone structure. Note that the base portals have the numbers one and two.
}
\examples{
g <- fishbone(8)
plot(g)
}
