% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlan.R
\name{getPlan}
\alias{getPlan}
\title{getPlan}
\usage{
getPlan(g, names = NULL, quiet = FALSE, wait = -1)
}
\arguments{
\item{g}{a web with  way(s) and link plan(s)}

\item{names}{character vector: names of portals (default: use numbers)}

\item{quiet}{logical: print ways and plans (default: FALSE)}

\item{wait}{logical: how long to wait in seconds between agent moves (default: -1)}
}
\value{
a list with ways &link plans, required items (keys, Softbank ultralinks)
}
\description{
After the way(s) and the link plan(s) are created they are shown, printed and returned invisibly- 
If \code{quiet=TRUE} then no printout will be generated. If \code{wait=0} then the agents move will not be plotted. 
A negative value for \code{wait} forces a keypress to continue after each agent move and a positive value makes a break of \code{wait}
seconds after each agent move.
}
\examples{
set.seed(0)
g  <- homogeneous(2)
x  <- cbind(runif(10), runif(10))
gi <- web(g, x)
g1 <- optimizeWeb(gi)
g2 <- linkPlan1(g1)
plan <- getPlan(g2, wait=0)
plan$items
plan[[1]]
}
