% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneous.R
\name{homogeneous}
\alias{homogeneous}
\title{homogeneous}
\usage{
homogeneous(n)
}
\arguments{
\item{n}{numeric: number of layers}
}
\value{
a web
}
\description{
Creates a homogeneous mutlilayer field (see \href{https://www.youtube.com/watch?v=OvxD-iMgVPM}{Michael Hartley's three part videos}).
}
\examples{
g <- homogeneous(2) # the most used homogeneous field
plot(g)
}
