% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leisepark.R
\docType{data}
\name{leisepark}
\alias{leisepark}
\title{Leisepark}
\format{An object of class \code{data.frame} with 43 rows and 5 columns.}
\usage{
data(leisepark)
}
\description{
The dataset contains the position (longitude, latitude) of 43 portals in the Leisepark in Berlin, Germany.
}
\examples{
data(leisepark)
head(leisepark)
plot(leisepark$lat, leisepark$lon)
text(leisepark$lat, leisepark$lon, labels=leisepark$shortname)
}
\keyword{datasets}
