% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPlan1.R
\name{linkPlan1}
\alias{linkPlan1}
\title{linkPlan1}
\usage{
linkPlan1(g, way = NULL, direction = NULL)
}
\arguments{
\item{g}{web}

\item{way}{numeric vector: a sequence of portals the agent should go (default: NULL)}

\item{direction}{numeric: angle main direction the agent will (default: NULL)}
}
\value{
a web with a way and link plan
}
\description{
Creates a way through the portals and a link plan for a single agent.
}
\details{
If \code{direction} is not set then the direction of the first principal component is as main direction choosen.
If \code{way} is set then first is checked if we can create a maximum field with the path. If yes you get the link plan bakc.
}
\examples{
set.seed(0)
g  <- homogeneous(2)
x  <- cbind(runif(10), runif(10))
gi <- web(g, x)
g1 <- optimizeWeb(gi)
g2 <- linkPlan1(g1)
\dontrun{
runPlan(g2)
}
}
