% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPlanN.R
\name{linkPlanN}
\alias{linkPlanN}
\title{linkPlanN}
\usage{
linkPlanN(g, agents = 1, start = NULL, dir = 0)
}
\arguments{
\item{g}{web: the current link plan}

\item{agents}{numeric: the number of agents (default: 1)}

\item{start}{numeric: number of start portal (default: the "most" inner portal)}

\item{dir}{numeric: direction to rotate sectors (default: 0)}
}
\value{
a web with a way and link plan
}
\description{
Tries to create a link plan for N agents, starting from the center of the target link structure. 
The portals are grouped in N sectors (one for each agent) which are accessed by distance from center
and linked back to all other portals.
}
\note{
Is is not clear if this approach leads always to the maximum field number!
}
