% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll2xy.R
\name{ll2xy}
\alias{ll2xy}
\title{ll2xy}
\usage{
ll2xy(lon, lat)
}
\arguments{
\item{lon}{numeric vector: longitudes of portal data}

\item{lat}{numeric vector: latitude of portal data}
}
\value{
a numeric matrix with xy coordinates.
}
\description{
Converts portal coordinates from (longitude, latitude) to xy using the \href{https://en.wikipedia.org/wiki/Universal_Transverse_Mercator_coordinate_system}{UTM projection}.
}
\details{
To install the package rgdal you may need to install under linux the appropriate system libraries. 
For details see in stackoverflow.com: \href{https://stackoverflow.com/questions/15248815/rgdal-package-installation}{rgdal package installation} and
\href{https://stackoverflow.com/questions/12141422/error-gdal-config-not-found}{Error: gdal-config not found}
}
\examples{
data(leisepark)
xy <- ll2xy(leisepark$lon, leisepark$lat)
g  <- homogeneous(2)
gi <- web(g, xy)
plot(gi)
}
