% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalie}
\usage{
normalize(mat)
}
\arguments{
\item{mat}{a numeric input matrix}
}
\value{
a sorted matrix
}
\description{
Sorts first in each row of a matrix the entries by value and secondly the rows by first, second etc. entry
}
\examples{
x <- matrix(round(5*runif(15)), ncol=3)
normalize(x)
}
