% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeWeb.R
\name{optimizeWeb}
\alias{optimizeWeb}
\title{optimizeWeb}
\usage{
optimizeWeb(gi, maxit = 1e+05, plot = 5, maxt = Inf)
}
\arguments{
\item{gi}{a web}

\item{maxit}{numeric: maximal number of iterations (default: 100000)}

\item{plot}{numeric: minimal time (in seconds) between consecutive plots (default: 5)}

\item{maxt}{numeric: maximal time (in seconds) the optimization runs (default: Inf)}
}
\value{
a web
}
\description{
Optimizes
}
\examples{
xy <- cbind(runif(20), runif(20))
g  <- fishbone(8)
gf <- web(g, xy)
gs <- optimizeWeb(gf)
}
