% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.web.R
\name{plot.web}
\alias{plot.web}
\title{plot.web}
\usage{
\method{plot}{web}(x, ...)
}
\arguments{
\item{x}{a web}

\item{...}{further parameters (unused)}
}
\description{
plot.web
}
\examples{
xy <- cbind(runif(30), runif(30))
g  <- fishbone(8)
plot(g)
gf <- web(g, xy)
plot(gf)
}
