% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsInConvexSet.R
\name{pointsInConvexSet}
\alias{pointsInConvexSet}
\title{pointsInConvexSet}
\usage{
pointsInConvexSet(px, set)
}
\arguments{
\item{px}{numeric matrix: portal position}

\item{set}{numeric vector: index of portals which form a convex set}
}
\value{
logical vector with \code{TRUE} if point is in the set otherwise false
}
\description{
Given the point indices in \code{set} form a convex set, it is checked 
for each point that if it is inside the convex set or not.
}
\examples{
set.seed(1)
x <- cbind(runif(10), runif(10))
inside <- pointsInConvexSet(x, c(1,7,10))
plot(x, col=inside+1, asp=TRUE, pch=19)
polygon(x[c(1,7,10),], border="red")
}
