% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointsInTrianglesC}
\alias{pointsInTrianglesC}
\title{Compute the number of points in trangles.}
\usage{
pointsInTrianglesC(vertices, faces)
}
\arguments{
\item{vertices}{numeric matrix with two columns: xy coordinates}

\item{faces}{numeric matrix with three columns: index of points which form the triangles}
}
\value{
a vector with number of points in each triangle
}
\description{
Compute the number of points in trangles.
}
\references{
\url{https://stackoverflow.com/questions/2049582/how-to-determine-if-a-point-is-in-a-2d-triangle}
}
