% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.web.R
\name{print.web}
\alias{print.web}
\title{print.web}
\usage{
\method{print}{web}(x, ...)
}
\arguments{
\item{x}{a web}

\item{...}{further parameters (unused)}
}
\description{
print.web
}
\examples{
xy <- cbind(runif(30), runif(30))
g  <- fishbone(8)
g
gf <- web(g, xy)
gf
}
