% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.web.R
\name{rev.web}
\alias{rev.web}
\title{rev}
\usage{
\method{rev}{web}(x)
}
\arguments{
\item{x}{a web}
}
\value{
a vector of portal indices
}
\description{
Extracts the mapping from the real portals to the portals of the target linking structure
}
\examples{
data(leisepark)
xy <- ll2xy(leisepark$lon, leisepark$lat)
g <- fishbone(8)
g0 <- web(g, xy)
rev(g0)
}
