% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.web.R
\name{summary.web}
\alias{summary.web}
\title{summary.web}
\usage{
\method{summary}{web}(object, ...)
}
\arguments{
\item{object}{a web}

\item{...}{further parameters (unused)}
}
\description{
Prints a summary of a web.
}
\examples{
xy <- cbind(runif(30), runif(30))
g  <- fishbone(8)
summary(g)
gf <- web(g, xy)
summary(gf)
}
