% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{swap}
\alias{swap}
\title{swap}
\usage{
swap(g, i, j)
}
\arguments{
\item{g}{a web}

\item{i}{numeric: index of first portal to be swapped}

\item{j}{numeric: index of second portal to be swapped}
}
\value{
the modified web
}
\description{
Swaps the given structure between portals. Mainly for use in \link{optimizeWeb}.
}
\examples{
par(mfrow=c(1,2))
g <- fishbone(8)
plot(g)
gs <- swap(g, 1, 3)
plot(gs)
}
