% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapi.R
\name{swapi}
\alias{swapi}
\title{swapi}
\usage{
swapi(g, ...)
}
\arguments{
\item{g}{a web}

\item{...}{further parameters for plotting, see \code{\link{plot.web}}}
}
\value{
the modified web
}
\description{
Swaps interactively the given structure between portals for web optimization by hand. 
To finish press Esc or choose two portals which do no belong to the given linking structure.
}
\examples{
\dontrun{
xy <- cbind(runif(20), runif(20))
g  <- fishbone(8)
gf <- web(g, xy)
gs <- swapi(gf)
summary(gs)
}
}
