% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\alias{web.default}
\title{web}
\usage{
web(g, ...)

\method{web}{default}(g, px, k = 10, rot = 8, ...)
}
\arguments{
\item{g}{web: a given linking structure}

\item{...}{further parameters (unused)}

\item{px}{numeric matrix: xy-coordinates of portals}

\item{k}{numeric: number of neighbours (default: 10)}

\item{rot}{numeric: number of rotations to find a good starting position (default: 8)}
}
\value{
web with starting configuration

\code{NULL}
}
\description{
Creates an initial configuration given by g on a set of portals.
}
\examples{
xy <- cbind(runif(30), runif(30))
g  <- fishbone(8)
gf <- web(g, xy)
plot(gf)
}
